demos/grasp_data_generator/requirements.txt
chainer
chainer_mask_rcnn
chainercv
chainermn
cupy
cython
fcn
imgaug
matplotlib
numpy
pyyaml
scipy
scikit-image
six
